+++ ios-base/Extensions/ColorExtension.swift
@@ -8,7 +8,6 @@
 
 import UIKit
 
-// swiftlint:disable force_unwrapping
 extension UIColor {
   
   static let buttonBackground = UIColor(named: "buttonBackground")!
