+++ ios-base/Helpers/UIConstants.swift
@@ -1,0 +1,28 @@
+//
+//  UIConstants.swift
+//  ios-base
+//
+//  Created by Karen Stoletniy on 13/7/21.
+//  Copyright © 2021 Rootstrap Inc. All rights reserved.
+//
+
+import Foundation
+import UIKit
+
+struct UI {
+  enum Defaults {
+    static let margin: CGFloat = 32
+  }
+  
+  enum ViewController {
+    static let topMargin: CGFloat = 72
+  }
+  
+  enum Button {
+    static let cornerRadius: CGFloat = 22.0
+    static let height: CGFloat = 45.0
+    static let width: CGFloat = 200.0
+    static let spacing: CGFloat = 20.0
+  }
+}
+
